biodv: in type Point now altitude and deph are now uint (as they should be in meters); driver/gbif: return collection event is updated to return uints on depth and elevation; records: records read, as well as changed, now set altitude and depth as uint; cmd/biodv: in command rec.value, altitude and depth are treated as uint

+++ geography.go
@@ -36,8 +36,8 @@ func (c CollectionEvent) CountryName() string {
 type Point struct {
 	Lat         float64
 	Lon         float64
-	Altitude    float64
-	Depth       float64
+	Altitude    uint   // in meters
+	Depth       uint   // in meters
 	Source      string // source of the reference
 	Uncertainty uint   // georeference uncertainty in meters
 	Validation  string // source of the validation
